Added getRenderingWidth
This seperates out the explicit width (as set by width) and the
calculated width (used for rendering). This allows the detection of
groups being their default size or an explicit size. It also means width
and height are now symetric. I guess there are other advantages to this
as well.