load supports multithreading via workers
Basically, it spins up hardwareConcurrency/2 threads which can then be
sent "orders". These orders are then processed and the output is then
sent back to the main thread and a promise resolves. Orders automaticaly
go to an unused worker (or wait until one is idle).

Some other changes were needed in load, though, including changing the
URLs to automatically generate absolute URLs and a bit of reordering of
code to support synchronous imports.