Changed how text is rendered
Instead of having big functions or whatever, it uses two helper classes;
Location, which is the location of the "painting cursor" at a given
point and FormatBlock which represents a single "node" of a tree (like
DOM nodes, I guess?). Essentially, each FormatBlock is in charge of
painting text, and different formats (italics, bold...) can subclass it
to change the formatting. In the future, _processText can build a tree,
but right now it only creates one root node.

In addition, it now uses getRenderingWidth/Height rather than setting
and getting the width/height directly. So hurray for that.

However, now you must specify the size of text explictly, rather than
setting the width and it working automatically. Which is probably better
design anyway.